Add xBee_driver unit tests module - RX test
Those files are quite new in vive tracker project development. They can be used to unit test the xBee driver if something doesn't work in the future with the xBee communication.

Only the RX unit test is implemented. The TX must be working properly before. The TX test must be implemented ASAP.