Arsenal and Tonnerre protocols implementation
A partial implementation of Arsenal and Tonnerre's protocol. Trackers are now able to communicate their positions (not implemented in main program yet) and some basic messages to help robots to cooperate. The acknowledgement system is not fully functionnal : retries are missing.

New feature : Vive trackers ID are now set with jumpers 5 (bit #0) and 6 (bit #1). Available IDs : from 0 to 3.

Some bugfixes related to protocols :
- USB_commands_manager didn't recognize end characters correctly if end character wasn't \n or space.
- xBee_driver get the right src_addr now (endianness correction).